+++ tests/testthat/test-plotly-parcats.R
@@ -5,25 +5,13 @@ expect_traces <- function(p, n.traces, name) {
   L
 }
 
-test_that("No cartesian axes are supplied to a splom chart", {
-  
-  p <- plot_ly(
-    type = 'splom',
-    dimensions = list(
-      list(values=c(1,2,3), label="A"),
-      list(values=c(2,5,6), label="B")
-    )
-  )
-  
-})
-
 test_that("values property has a class of AsIs", {
   p <- plot_ly(
-    type = "splom",
     dimensions = list(
-      list(values = 1, label = "A"),
-      list(values = 2, label = "B")
-    )
+      list(values = "A"),
+      list(values = "B")
+    ),
+    type = "parcats"
   )
   l <- expect_traces(p, 1, "parcats-data-array")
   tr <- l$data[[1]]$dimensions